requirements/requirements.txt
opencv-python==4.5.5.64
scikit-image
onnx
onnxruntime-gpu
GPUtil